oid85-01-04 add tinkoffservice

+++ Oid85.FinMarket/Oid85.FinMarket.Domain/Models/Candle.cs
@@ -28,7 +28,7 @@
         /// <summary>
         /// Объем
         /// </summary>
-        public int Volume { get; set; }
+        public long Volume { get; set; }
 
         /// <summary>
         /// Дата
